+++ packages/nextjs/app/(landing)/page.tsx
@@ -1,0 +1,3 @@
+export default function HomePage() {
+  return <>this is landingpage</>;
+}
