+++ packages/nextjs/app/(main)/ai-assistant/layout.tsx
