+++ packages/nextjs/app/(main)/api/price/[symbol]/route.ts
