+++ packages/nextjs/app/(main)/debug/_components/DebugContracts.tsx
@@ -2,7 +2,7 @@
 
 import { useEffect } from "react";
 import { useLocalStorage } from "usehooks-ts";
-import { ContractUI } from "~~/app/debug/_components/contract";
+import { ContractUI } from "~~/app/(main)/debug/_components/contract";
 import { ContractName } from "~~/utils/scaffold-stark/contract";
 import { getAllContracts } from "~~/utils/scaffold-stark/contractsData";
