+++ packages/nextjs/app/(main)/debug/_components/contract/ContractVariables.tsx
