+++ packages/nextjs/app/(main)/debug/_components/contract/DisplayVariable.tsx
