+++ packages/nextjs/app/(main)/debug/_components/contract/InheritanceTooltip.tsx
