+++ packages/nextjs/app/(main)/debug/_components/contract/Struct.tsx
