+++ packages/nextjs/app/(main)/debug/_components/contract/utilsDisplay.tsx
