+++ packages/nextjs/app/(main)/layout.tsx
@@ -1,0 +1,26 @@
+import type { Metadata } from "next";
+import { ScaffoldStarkAppWithProviders } from "~~/components/ScaffoldStarkAppWithProviders";
+import "~~/styles/globals.css";
+import { ThemeProvider } from "~~/components/ThemeProvider";
+
+export const metadata: Metadata = {
+  title: "Scaffold-Stark",
+  description: "Fast track your starknet journey",
+  icons: "/logo.ico",
+};
+
+const ScaffoldStarkApp = ({ children }: { children: React.ReactNode }) => {
+  return (
+    <html suppressHydrationWarning>
+      <body>
+        <ThemeProvider>
+          <ScaffoldStarkAppWithProviders>
+            {children}
+          </ScaffoldStarkAppWithProviders>
+        </ThemeProvider>
+      </body>
+    </html>
+  );
+};
+
+export default ScaffoldStarkApp;
