+++ packages/nextjs/app/(main)/transaction/receive/_component/ReceiveToken.tsx
