--- packages/nextjs/app/page.tsx
@@ -1,11 +1,0 @@
-"use client";
-import type { NextPage } from "next";
-import { useAccount } from "~~/hooks/useAccount";
-
-const Home: NextPage = () => {
-  const connectedAddress = useAccount();
-
-  return <></>;
-};
-
-export default Home;
