+++ packages/nextjs/next.config.mjs
@@ -12,15 +12,6 @@ const nextConfig = {
     config.externals.push("pino-pretty", "lokijs", "encoding");
     return config;
   },
-  async redirects() {
-    return [
-      {
-        source: "/",
-        destination: "/overview",
-        permanent: true,
-      },
-    ];
-  },
   output: "standalone",
 };
